@@ -5,7 +5,7 @@
     "license": "MIT",
     "type": "project",
     "require": {
-        "php": ">=7.3.0",
+        "php": ">=7.4.0",
         "laravel/lumen-framework": "^8.0",
         "vlucas/phpdotenv": "^5.2",
         "league/flysystem": "^1.0"
