@@ -1,7 +1,7 @@
-<?xml version="1.0" encoding="UTF-8"?>
+<?php echo('<?xml version="1.0" encoding="UTF-8"?>'); ?>
 <root>
 <?php 
 foreach($storedValue as $key=>$value){
-    echo("<$key>".htmlentities($value)."</$key>"); 
+    echo("<$key>" . $value . "</$key>");
 } ?>
 </root>
